+++ frontend/src/web/app/Root.js
@@ -4,7 +4,8 @@ import { Provider as Redux } from 'react-redux';
 import App from './App';
 
 type Props = {
-  store: Object;
+  store: Object,
+  history: Object,
 };
 
 // Use ES6 class to ensure hot reload works for stateless components.
@@ -13,10 +14,10 @@ class Root extends React.Component {
   props: Props;
 
   render() {
-    const { store } = this.props;
+    const { store, history } = this.props;
     return (
       <Redux store={store}>
-        <App />
+        <App history={history} />
       </Redux>
     );
   }