+++ frontend/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2017-03-30
 ### Added
 - Add application global fatal error message box with APP_FATAL_ERROR action.
+- Add react-router-redux package to support saving routing in redux store.
 
 ### Fixed
 - Fix component recreate issue caused by HOC component.